refactor: move ModelUtils in custom folder
